% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{evaluate}
\usage{
evaluate(
  network,
  reference_set,
  output_directory,
  network_name,
  edges = c(512, 1024, 2048, 4096, 8192, 16384, 32768, 65536),
  num_possible_TFs = 0,
  permutations = 3,
  get_sum = TRUE,
  get_percent = FALSE,
  get_mean = FALSE,
  get_median = FALSE,
  get_annotation_overlap = FALSE,
  get_size = TRUE,
  plot = TRUE
)
}
\arguments{
\item{network}{a file containing the network to create subsets from. The file should
be tab-separated with three columns: source node, target node, edge score}

\item{reference_set}{path to the set of all genes possibly included in the network. Must be a .txt
file containing exactly one column of the genes that could possibly appear in the network.}

\item{output_directory}{path to the directory in which to store the generated network subsets,
ORA summaries, and plots}

\item{network_name}{short name for the network - used in file naming so may not contain spaces}

\item{edges}{list of total numbers of edges or average edges per TF to include in each subset}

\item{num_possible_TFs}{if set to a number > 0, the elements of 'edges' will first be
multiplied by this number to get the number of edges for each subset}

\item{permutations}{the number of randomly permuted networks to create and run ORA}

\item{get_sum}{bool whether to get the 'sum' metric, which is the sum of the negative
log base 10 of the p-value for the top term of each TF minus 3 times the total
number of TFs.}

\item{get_percent}{bool whether to get the 'percent' metric, which is the
percent of TFs with at least one GO term with a FDR < 0.05}

\item{get_mean}{bool whether to get the 'mean' metric, which is the mean negative
log base 10 of the p-value for the top term of each TF}

\item{get_median}{bool whether to get the 'median' metric, which is the median negative
log base 10 of the p-value for the top term of each TF}

\item{get_annotation_overlap}{bool whether to get the 'annotation_overlap' metric,
which is the percent of TFs that are annotated to a GO term for which their
target genes are enriched}

\item{get_size}{bool whether to get the 'size' metric, which is the number of
TFs in the network subset that have more than one target gene with GO annotations}

\item{plot}{bool whether to make a pdf containing plots of the calculated metrics}
}
\description{
Perform the whole GOeval pipeline on one network. Given one network, this
function will first make subsets with subset_network, then run ORA with
webgestalt_network, followed by get_metrics and plot_metrics to plot summary
statistics.
The input file should be tab-separated with three columns: source node (e.g. transcription factor),
target node (e.g. the regulated gene), and edge score. All genes should be written as Ensembl gene IDs.
}
