% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{get_annotation_overlap}
\alias{get_annotation_overlap}
\title{calculate the percent of TFs that are annotated to a GO term for which their
target genes are enriched}
\usage{
get_annotation_overlap(terms, organism, go_ann, go_reg)
}
\arguments{
\item{terms}{a data.frame obtained from calling get_terms on a folder of the
summaries output by webgestalt_network}

\item{organism}{a string specifying the organism that the data is from, e.g.
"hsapiens" or "scerevisiae"}

\item{go_ann}{a data.frame of annotations of genes to GO terms. Obtain with
WebGestaltR::loadGeneSet.}

\item{go_reg}{a data.frame of the regulatory relationships between GO terms.
Obtain with ontologyIndex::get_ontology.}
}
\description{
calculate the percent of TFs that are annotated to a GO term for which their
target genes are enriched
}
