% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_terms.R
\name{get_terms}
\alias{get_terms}
\title{get enrichment data for the top n terms from each tf summary file in a folder
output by the webgestalt_network function}
\usage{
get_terms(path, n)
}
\arguments{
\item{path}{a folder containing .csv files of the enrichment summaries output
by webgestalt_network}

\item{n}{the number of terms from each summary to get data for}
}
\description{
get enrichment data for the top n terms from each tf summary file in a folder
output by the webgestalt_network function
}
