% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_network.R
\name{subset_network}
\alias{subset_network}
\title{subset_network}
\usage{
subset_network(input_file, output_folder, name, edges, perTF)
}
\arguments{
\item{input_file}{a folder containing .csv files of the enrichment summaries output
by webgestalt_network}

\item{output_folder}{name of the folder in which to place the created network subset files}

\item{name}{base name for the created network subset files}

\item{edges}{list of numbers of edges or average edges per TF to include in each subset}

\item{perTF}{boolean TRUE if the elements of "edges" represent average edges per TF
in the subsets and FALSE if they represent the total number of edges}
}
\description{
From a network with edge scores, sorts the network in descending order of score, and
creates subsets using the top given numbers of edges.
The input file should be tab-separated with three columns: TFs, regulated genes, scores.
The output files will be tab-separated with two columns: TFs, regulated genes.
}
