% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webgestalt_network.R
\name{webgestalt_network}
\alias{webgestalt_network}
\title{given a network or network subset, webgestalt_network will generate ORA results
for this network and a given number of permutations
NOTE: running with a previously used network_name will cause overwriting
NOTE: if using the same output_directory for a different network than before,
you must first delete the contents of "/webgestalt_work/network_name/"}
\usage{
webgestalt_network(
  network_path,
  reference_set,
  output_directory,
  network_name,
  organism = "hsapiens",
  database = "geneontology_Biological_Process_noRedundant",
  permutations = 10
)
}
\arguments{
\item{network_path}{path to the network or network subset to evaluate}

\item{reference_set}{path to the set of all genes possibly included in the network}

\item{output_directory}{path to the folder in which output from all networks subset from the same original network should be stored}

\item{network_name}{the name of the folder to store the results within the output_directory}

\item{organism}{human: "hsapiens"; yeast: "scerevisiae"}

\item{database}{the gene set database to search for enrichment - see options with WebGestaltR::listGeneSet()}

\item{permutations}{the number of randomly permuted networks to create and run ORA}
}
\description{
given a network or network subset, webgestalt_network will generate ORA results
for this network and a given number of permutations
NOTE: running with a previously used network_name will cause overwriting
NOTE: if using the same output_directory for a different network than before,
you must first delete the contents of "/webgestalt_work/network_name/"
}
