% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{get_metrics}
\usage{
get_metrics(
  directory,
  organism = "hsapiens",
  get_sum = TRUE,
  get_percent = FALSE,
  get_mean = FALSE,
  get_median = FALSE,
  get_annotation_overlap = FALSE,
  get_size = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{directory}{a directory containing the webgestalt_network output directory
for all networks of interest}

\item{organism}{a string specifying the organism that the data is from, e.g.
"hsapiens" or "scerevisiae". Only specify if get_annotation_overlap = TRUE.}

\item{get_sum}{bool whether to get the 'sum' metric, which is the sum of the negative
log base 10 of the p-value for the top term of each TF minus 3 times the total
number of TFs.}

\item{get_percent}{bool whether to get the 'percent' metric, which is the
percent of TFs with at least one GO term with a FDR < 0.05}

\item{get_mean}{bool whether to get the 'mean' metric, which is the mean negative
log base 10 of the p-value for the top term of each TF}

\item{get_median}{bool whether to get the 'median' metric, which is the median negative
log base 10 of the p-value for the top term of each TF}

\item{get_annotation_overlap}{bool whether to get the 'annotation_overlap' metric,
which is the percent of TFs that are annotated to a GO term for which their
target genes are enriched}

\item{get_size}{bool whether to get the 'size' metric, which is the number of
TFs in the network subset that have more than one target gene with GO annotations}

\item{parallel}{bool whether to get the metrics for each network in the directory
in parallel or sequentially - use with caution, as this has not been adequately tested}
}
\description{
Get a data.frame that contains specified metrics for all networks that have a
subdirectory in the provided path including both the real and permuted networks.
The path should contain only directories created by webgestalt_network.
}
