% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{get_network_metrics}
\alias{get_network_metrics}
\title{helper function to compute the desired metrics from a dataframe of the top
terms for each TF in a single network}
\usage{
get_network_metrics(
  full_terms,
  network_size,
  organism,
  get_sum,
  get_percent,
  get_mean,
  get_median,
  get_annotation_overlap,
  go_ann = NULL,
  go_reg = NULL
)
}
\arguments{
\item{full_terms}{a data.frame containing all the summary results for a network.
Use get_terms to obtain.}

\item{network_size}{the number of TFs in the network. May be different from
the number of TFs that have summary files.}

\item{organism}{a string specifying the organism that the data is from, e.g.
"hsapiens" or "scerevisiae". Only specify if get_annotation_overlap = TRUE.}

\item{get_sum}{bool whether to get the 'sum' metric, which is the sum of the negative
log base 10 of the p-value for the top term of each TF minus 3 times the total
number of TFs.}

\item{get_percent}{bool whether to get the 'percent' metric, which is the
percent of TFs with at least one GO term with a FDR < 0.05}

\item{get_mean}{bool whether to get the 'mean' metric, which is the mean negative
log base 10 of the p-value for the top term of each TF}

\item{get_median}{bool whether to get the 'median' metric, which is the median negative
log base 10 of the p-value for the top term of each TF}

\item{get_annotation_overlap}{bool whether to get the 'annotation_overlap' metric,
which is the percent of TFs that are annotated to a GO term for which their
target genes are enriched}

\item{go_ann}{a data.frame of annotations of genes to GO terms. Obtain with
WebGestaltR::loadGeneSet.}

\item{go_reg}{a data.frame of the regulatory relationships between GO terms.
Obtain with ontologyIndex::get_ontology.}
}
\description{
helper function to compute the desired metrics from a dataframe of the top
terms for each TF in a single network
}
