% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metrics.R
\name{plot_metrics}
\alias{plot_metrics}
\title{plot_metrics}
\usage{
plot_metrics(
  metric_dfs_by_net,
  title_text,
  subtitle_text,
  perTF,
  sum = TRUE,
  percent = FALSE,
  mean = FALSE,
  median = FALSE,
  annotation_overlap = FALSE,
  size = TRUE
)
}
\arguments{
\item{metric_dfs_by_net}{a list of outputs of get_metrics for each network to be plotted}

\item{title_text}{text for the title of each plot; generally the names of the networks}

\item{subtitle_text}{text for the subtitle of each plot}

\item{perTF}{bool whether the network subset sizes were specified as average target
genes per TF; changes the x-axis label}

\item{sum}{bool whether to plot the 'sum' metric, which is the sum of the negative
log base 10 of the p-value for the top term of each TF minus 3 times the total
number of TFs.}

\item{percent}{bool whether to plot the 'percent' metric, which is the
percent of TFs with at least one GO term with a FDR < 0.05}

\item{mean}{bool whether to plot the 'mean' metric, which is the mean negative
log base 10 of the p-value for the top term of each TF}

\item{median}{bool whether to plot the 'median' metric, which is the median negative
log base 10 of the p-value for the top term of each TF}

\item{annotation_overlap}{bool whether to plot the 'annotation_overlap' metric,
which is the percent of TFs that are annotated to a GO term for which their
target genes are enriched}
}
\description{
Plot the summary metrics from get_metrics across network subset size for one or more networks.
}
