% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webgestalt_network.R
\name{webgestalt_network}
\alias{webgestalt_network}
\title{webgestalt_network}
\usage{
webgestalt_network(
  network_path,
  reference_set,
  output_directory,
  network_name,
  organism = "hsapiens",
  database = "geneontology_Biological_Process_noRedundant",
  permutations = 10
)
}
\arguments{
\item{network_path}{path to the network or network subset to evaluate. Must be a tab-separated
file where the first column is the source nodes (TFs) and the second column is the target nodes (regulated genes).}

\item{reference_set}{path to the set of all genes possibly included in the network. Must be a .txt
file containing exactly one column of the genes that could possibly appear in the network.}

\item{output_directory}{path to the folder in which output from all networks subset from the same original network should be stored}

\item{network_name}{the name of the folder to store the results within the output_directory.
It should be in the same format as the files output by subset_network: "\{name\}_\{# of edges\}" i.e. "example_8".}

\item{organism}{human: "hsapiens"; yeast: "scerevisiae"}

\item{database}{the gene set database to search for enrichment - see options with WebGestaltR::listGeneSet()}

\item{permutations}{the number of randomly permuted networks to create and run ORA}
}
\description{
Given a network or network subset, webgestalt_network will generate ORA results
for this network and a given number of permutations.
The input network file should be a tab-separated .tsv where the first column is the
source nodes (TFs) and the second column is the target nodes (regulated genes).
All genes must be written as their Ensembl gene ID.
}
